feat: move new Popover from lab
